06-Command-and-Control-C2-Systems/StegoC2/requirements.txt
Pillow==10.2.0
requests==2.31.0
flask==3.0.2
pyyaml==6.0.1
